Add files via upload
A function isLeap that takes as an input parameter an integer representing a year, and returns True if that year is a leap year, and False otherwise. Recall the rules for leap years:
Years divisible by 4, but not divisible by 100, are leap years. So, 1984 and 2004 are leap years
Years divisible by 100, but not divisible by 400, are not leap years. So, 1900 is not a leap year
Years divisible by 400 are leap years. So, 2000 is a leap year
>>> isLeap(1984)
True
>>> isLeap(2000)
True
A function daysSinceYearStart that takes as input parameters two integers representing a month and a year, and returns the number of days before the 1st day of that month, in that year. So if you call this function with 1 and 2000, it will return 0 (since there are 0 days before Jan 1 in the year 2000. If you call this function with 4 and 1900, it will return 90 (since there are 90 days before April 1st in the year 1900), and if you call it with 4 and 200, it will return 91 (remember those leap years!!) This function should use your isLeap function.
>>> daysSinceYearStart(1,2000)
0
>>> daysSinceYearStart(4,1900)
90
>>> daysSinceYearStart(4,200)
90
A function dateDifference that takes as input 6 parameters: month1, day1, year1 and month2, day2, year2, and returns the difference between those two dates. You may assume that year1 and year2 are the same (unless you do the extra credit!). You should not assume that the first date comes before the second date!
>>> dateDifference(1,15,2015, 3, 17, 2015)
61
>>> dateDifference(10, 31, 2015, 10, 15, 2015)
16

From David Galles Intro to Computer Science 1, fall 2015 lab3