Lab discription
squareRoot(n) To calculate the square root of a number, you should use Newton's method. The square root of a number n can be determined by repeated calculation using the following formula:
NextGuess = 0.5 * (CurrentGuess + n / CurrentGuess)
For instance, if we are trying to calculate the square root of 7, and we start with the initial guess of 1, we get:

NextGuess = 0.5 * (1 + 7 / 1) = 4
Our current guess is 4, we go again:

NextGuess = 0.5 * (4 + 7 / 4) = 2.875
Our current guess is 2.875, so we go again:

NextGuess = 0.5 * (2.875 + 7 / 2.875) = 2.654891304347826
We can go again:

NextGuess = 0.5 * (2.654891304347826 + 7 / 2.654891304347826) = 2.64576704419029
You can see that we are getting closer and closer to the actual value of the square root. The longer we go, the closer we get to the actual value. When do we stop? When the difference between the guesses gets small enough. A difference of 0.0001 is a good stopping place. (So, when the absolute value of the CurrentGuess - NextGuess < 0.0001)

>>> x = squareRoot(2)
>>> print(x)
1.4142135623746899
>>> print(x * x)
2.0000000000045106
Your numbers do not need to be exactly the same as the examples!

log(n).  (Base 2).  Recall that log(x) = y if and only if 2y = x.  So, log(2) = 1, log(4) = 2, log(8) = 3, log(1024) = 10. Calculating the floating point log can be difficult, so we are going to calculate integer log. Just as // returns the result of division after we throw away the remainder, our log will calculate the result of taking the log after we throw away the remainder. Since log(x) = y if and only if 2y = x, then log(x) = the number of times we need to divide x by 2 (using integer division, //) until we get to 1.
log(1) = 0, since we need to divide 1 by 2 zero times to get to 1
log(2): 2 // 2 = 1, so log(2) = 1
log(4): 4 // 2 = 2, 2 // 2 = 1. We needed to divide 4 by 2 twice to get to 1, so log(4) = 2
log(64): 64 // 2 = 32, 32 // 2 = 16, 16 // 2 = 8, 8 // 2 = 4, 4 // 2 = 2, 2 // 2 = 1. We needed to divide 64 by 2 six times to get to 1, so log(64) = 6
log(57): 57 // 2 = 28, 28 // 2 = 14, 14 // 2 = 7, 7 // 2 = 3, 3 // 2 = 1, so log(57) = 5
>>> log(1)
0
>>> log(2)
1
>>> log(3)
1
>>> log(4)
2
>>> log(5)
2
>>> log(6)
2
>>> log(7)
2
>>> log(8)
3
>>> log(9)
3
>>> log(57)
5
>>> log(64)
6
From David Galles Intro to Computer Science 1, Fall 2015 lab4