lab description
findFirstIndex(L, x)
Returns the first index where x appears in L (using ==), or -1 if no such index exists. Note that you should not use the .index method on lists, since that will raise an error if the element is not in the list! (If you have no idea what the /index method is, since we did not cover it in class, then you can ignore this note)

>>> L = [1,2,3,2,5,2,6]
>>> findFirstIndex(L,2)
1
>>> findFirstIndex(L,6)
6
>>> findFirstIndex(L,99)
-1
>>> L2 = ["cat", "dog", "mouse", "cat"]
>>> findFirstIndex(L2, "cat")
0
>>> findFirstIndex(L2, "mouse")
2
findAllIndices(L, x)
Return a list of all indices that are equal to x, using ==. This list could be empty.

>>> L = [1,2,3,2,5,2,6]
>>> findAllIndices(L,2)
[1, 3, 5]
>>> findAllIndices(L,6)
[6]
>>> findAllIndices(L,99)
[]
>>> L2 = ["cat", "dog", "mouse", "cat"]
>>> findAllIndices(L2,"cat")
[0, 3]
removeLastCopy(L,x)
Return a copy of the list with the last element removed. If the element is not in the list, return an unchanged copy of the list.

>>> L = [1,2,3,4,3,5]
>>> L2 = removeLastCopy(L, 3)
>>> L
[1, 2, 3, 4, 3, 5]
>>> L2
[1, 2, 3, 4, 5]
>>> L3 = removeLastCopy(L,99)
>>> L3 = removeLastCopy(L,99)
>>> L3
[1, 2, 3, 4, 3, 5]
>>> L3[2] = 99
>>> L3
[1, 2, 99, 4, 3, 5]
>>> L
[1, 2, 3, 4, 3, 5]
From David Galles Intro to Computer Science 1, Fall 2015